n8n-popularity-system/n8n-popularity-system/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
requests
google-api-python-client
pytrends
